basic spring boot connection
Spring Cloud Vault uses by default the generic secret backend. In our
case we don’t need the generic backend. So we need to disable the
generic backend and enable the MySQL backend. All configuration needs to
be specified in the bootstrap configuration. For this example we use
bootstrap.yml in src/main/resources: